move file service in wrong directory to file service
